\name{ecoligenomeBNUM}
\alias{ecoligenomeBNUM}
\docType{data}
\title{ Environment for 'bnum' identifiers}
\description{
  An environment to associate Affymetrix probe set IDs with 'bnum' IDs
}
\usage{data(ecoligenomeBNUM)}
\format{
  The format is:
length 0 <environment>
}
\details{
  Escherichia coli genes are sometimes identified by 'bnum's. This
  identfier is typically a 'b' followed by digits.  
}
\source{
  
}
\references{
  
}
\keyword{datasets}
