\name{multiFun}
\alias{multiFun}
\alias{ecoligenomeMULTIFUN2GO}
\docType{data}
\title{multiFun classification}
\description{
  The MultiFun classification scheme
}
\usage{
data(multiFun)
data(ecoligenomeMULTIFUN2GO)
}
\format{
These are environments.
}
\details{
  
}
\source{
\url{http://genprotec.mbl.edu/files/MultiFun.txt}
}
\examples{
## To be done...
}
\keyword{datasets}
